Add exercise 4.21

+++ part4/bloglist/backend/controllers/blogs.js
@@ -3,6 +3,14 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+const decodeTokenAndFindUser = async (token, response) => {
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  return await User.findById(decodedToken.id)
+}
+
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
   .find({}).populate('user', {username: 1, name: 1})
@@ -11,12 +19,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  const user = await decodeTokenAndFindUser(request.token, response)
 
   const blog = new Blog({
     title: body.title,
@@ -34,8 +37,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = await decodeTokenAndFindUser(request.token, response)
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog === null) {
+    response.status(404).end()
+  } else if (blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else
+  response.status(401).json({ error: 'token invalid' })
 })
 
 blogsRouter.put('/:id', async(request, response) => {
